Create directory structure
