+++ src/main/java/com/google/sps/placeGuide/PlaceGuide.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.sps.place;
+package com.google.sps.placeGuide;
 
 import org.jetbrains.annotations.Nullable;
